+++ AdventureBackpacks/Assets/Backpacks.cs
@@ -7,7 +7,6 @@ using AdventureBackpacks.Assets.Items;
 using AdventureBackpacks.Components;
 using AdventureBackpacks.Extensions;
 using BepInEx;
-using BepInEx.Bootstrap;
 using Vapok.Common.Abstractions;
 using Vapok.Common.Managers;
 using Vapok.Common.Managers.StatusEffects;
@@ -210,9 +209,13 @@ namespace AdventureBackpacks.Assets
                 return null;
             
             var backpackName = itemData.m_shared.m_name;
+            
             var backpackQuality = itemData.m_quality;
+            
             var statusEffects = new CustomSE(Enums.StatusEffects.Stats, $"SE_{backpackName}_{backpackQuality}");
+            
             var defaultStatusName = $"{backpackName} $vapok_mod_level {backpackQuality} $vapok_mod_effect";
+            
 
             if (backpack.ShowBackpackStatusEffect.Value)
             {
@@ -220,7 +223,7 @@ namespace AdventureBackpacks.Assets
                 statusEffects.Effect.m_startMessageType = MessageHud.MessageType.TopLeft;
                 statusEffects.Effect.m_startMessage = $"$vapok_mod_useful_backpack";
             }
-
+            
             var modifierList = new List<HitData.DamageModPair>();
             //Set Armor Default
             itemData.m_shared.m_armor = itemData.m_shared.m_armorPerLevel * backpackQuality;
@@ -244,20 +247,6 @@ namespace AdventureBackpacks.Assets
                 itemData.m_shared.m_setSize = 0;
                 itemData.m_shared.m_setStatusEffect = null;
             }
-
-            if (Chainloader.PluginInfos.ContainsKey("com.chebgonaz.ChebsNecromancy"))
-            {
-                var necroEffect = EffectsFactory.EffectList[BackpackEffect.NecromancyArmor];
-                //Apply Necromancy Armor Set if configured.
-                if (necroEffect.HasActiveStatusEffect(itemData, out var necroSetEffect))
-                {
-                    itemData.m_shared.m_setStatusEffect = necroSetEffect;
-                }
-                else
-                {
-                    itemData.m_shared.m_setStatusEffect = null;
-                }
-            }
             
             backpack.UpdateStatusEffects(backpackQuality, statusEffects, modifierList, itemData);
             
@@ -269,6 +258,7 @@ namespace AdventureBackpacks.Assets
             }
 
             itemData.AddSEToItem(statusEffects);
+            
             return statusEffects;
         }
     }